Migration to UI class
- finished migrating old UI controller script to new class module
- removed ui-controller.js
- migrated various scattered functions to HelperFunctions.js
- cleaned up unnecessary class variables
