+++ index.html
@@ -342,6 +342,5 @@
 <script src="main.js" type="module"></script>
 <script src="map.js" type="module"></script>
 <script src="atlas.js" type="module"></script>
-<script src="ui-controller.js"></script>
 
 </html>
