+++ main.js
@@ -1,11 +1,6 @@
-import { ROUND, CHOSEN_TIME } from './HelperFunctions.js';
+import { ROUND, CHOSEN_TIME, HOURS_TO_TIME_STRING } from './HelperFunctions.js';
 import DB from './classes/Database.js';
 import UI from './classes/UserInterface.js';
-
-//import SolarSystem from './classes/SolarSystem.js';
-//import CelestialBody from './classes/CelestialBody.js';
-//import Location from './classes/Location.js';
-//import Wormhole from './classes/Wormhole.js';
 
 window.SYSTEMS = Array();
 window.BODIES = Array();
@@ -14,11 +9,12 @@ window.WORMHOLES = Array();
 
 window.ACTIVE_LOCATION = null;
 window.SETTING_24HR = true;
-window.HOURS_TO_TIME_STRING = HOURS_TO_TIME_STRING;
-window.setText = setText;
 
 window.suppressReload = false;
 window.CHOSEN_TIME = 'now';
+
+
+
 
 // TEMPORARY SO SETTINGS.JS CAN USE IT
 window.setMapLocation = setMapLocation;
@@ -28,131 +24,14 @@ function setMapLocation(location) {
 // END TEMPORARY
 
 
-// FUNCTIONS
-function checkHash() {
-	let hash = window.location.hash;
-	if (hash === '') return;
 
-	let hashParts = hash.replace('#', '').replaceAll('_', ' ').split('@');
 
-	const locationName = hashParts[0];
-	const location = getLocationByName(locationName);
-
-	if (hashParts[1] !== undefined) {
-		setChosenTime(hashParts[1], true);
-	}
-
-	UI.setMapLocation(locationName);
-}
-
+// MAIN FUNCTIONS
 function update() {
-
 	if (window.LOCATIONS.length === 0) return;
+	const location = window.ACTIVE_LOCATION;
 
-	let location = window.ACTIVE_LOCATION;
-	let body = window.ACTIVE_LOCATION ? window.ACTIVE_LOCATION.PARENT : null;
-
-	const col = location.THEME_COLOR;
-	const colorMain = `rgb(${col.r}, ${col.g}, ${col.b})`;
-	const colorDark = `rgb(${col.r * 0.2}, ${col.g * 0.2}, ${col.b * 0.2})`;
-
-	document.querySelector(':root').style.setProperty('--theme-color', colorMain);
-	document.querySelector(':root').style.setProperty('--theme-color-dark', colorDark);
-
-	const bg = UI.el('selected-location-bg-image');
-	if (bg.backgroundColor !== colorMain) bg.backgroundColor = colorMain;
-
-	const bgImage = `url('${location.THEME_IMAGE}')`;
-	if (bg.style.backgroundImage !== bgImage) bg.style.backgroundImage = bgImage;
-
-
-	// MAIN LOCATION INFO
-	if (
-		location.ILLUMINATION_STATUS === 'Polar Day' ||
-		location.ILLUMINATION_STATUS === 'Polar Night' ||
-		location.LOCAL_TIME.toString() === 'NaN'
-	) {
-		UI.setText('local-time', location.ILLUMINATION_STATUS);
-	} else {
-		UI.setText('local-time', HOURS_TO_TIME_STRING(location.LOCAL_TIME / 60 / 60, false));
-	}
-	if (window.CHOSEN_TIME != 'now') {
-		UI.setText('chosen-time', CHOSEN_TIME().toLocaleString());
-		UI.setText('chosen-time-sublabel', 'local selected time');
-	} else {
-		UI.setText('chosen-time', '');
-		UI.setText('chosen-time-sublabel', '');
-	}
-	UI.setText('location-name', location.NAME);
-	UI.setText('location-body-name', location.PARENT.NAME);
-
-
-	// RISE/SET COUNTDOWNS
-	let nextRise = location.NEXT_STAR_RISE;
-	if (!nextRise) {
-		UI.setText('next-rise-countdown', '---');
-
-	} else {
-		nextRise = location.IS_STAR_RISING_NOW ? '- NOW -' : HOURS_TO_TIME_STRING(nextRise * 24, true, false);
-		UI.setText('next-rise-countdown', nextRise);
-	}
-
-	let nextSet = location.NEXT_STAR_SET;
-	if (!nextSet) {
-		UI.setText('next-set-countdown', '---');
-
-	} else {
-		nextSet = location.IS_STAR_SETTING_NOW ? '- NOW -' : HOURS_TO_TIME_STRING(nextSet * 24, true, false);
-		UI.setText('next-set-countdown', nextSet);
-	}
-
-
-	// RISE/SET LOCAL TIMES
-	if (!nextRise) {
-		UI.setText('local-rise-time', '---');
-	} else {
-		UI.setText('local-rise-time', HOURS_TO_TIME_STRING(location.LOCAL_STAR_RISE_TIME * 24, false, true));
-	}
-
-	if (!nextSet) {
-		UI.setText('local-set-time', '---');
-	} else {
-		UI.setText('local-set-time', HOURS_TO_TIME_STRING(location.LOCAL_STAR_SET_TIME * 24, false, true));
-	}
-
-
-	// RISE/SET REAL TIMES
-	let now = CHOSEN_TIME();
-	if (!nextRise) {
-		UI.setText('next-rise-time', '---');
-	} else {
-		let rise = now.setSeconds(now.getSeconds() + (location.NEXT_STAR_RISE * 86400));
-		// if (new Date(rise).getSeconds() === 59) {
-		// 	rise = new Date(rise + 1000);
-		// }
-		UI.setText('next-rise-time', DATE_TO_SHORT_TIME(new Date(rise)));
-	}
-
-	now = CHOSEN_TIME();
-	if (!nextSet) {
-		UI.setText('next-set-time', '---');
-	} else {
-		let set = now.setSeconds(now.getSeconds() + (location.NEXT_STAR_SET * 86400));
-		// if (new Date(set).getSeconds() === 59) {
-		// 	set = new Date(set.getTime() + 1000);
-		// }
-		UI.setText('next-set-time', DATE_TO_SHORT_TIME(new Date(set)));
-	}
-
-
-	// ILLUMINATION STATUS & IN-LORE CALENDAR DATE
-	let scDate = CHOSEN_TIME();
-	scDate.setFullYear(scDate.getFullYear() + 930);
-	let scDateString = scDate.toLocaleString('default', { year: 'numeric', month: 'long', day: 'numeric' });
-
-	UI.setText('illumination-status', location.ILLUMINATION_STATUS + '\r\n' + scDateString);
-
-
+	UI.update(location);
 	if (UI.showSettingsWindow) updateSettingsLocationTimes();
 	if (UI.showDebugWindow) updateDebugUI();
 }
@@ -273,20 +152,6 @@ function updateSettingsLocationTimes() {
 	}
 }
 
-// LEGACY IMPLEMENTATION; FIND REMAINING CALLS
-function setText(elementID, string) {
-	console.warn('Used old setText function!');
-
-	let el = (typeof elementID === 'string') ? document.getElementById(elementID) : elementID;
-
-	if (!el) {
-		console.error('Element not found:', elementID);
-		return null;
-	}
-
-	if (el.textContent !== string) el.textContent = string;
-}
-
 function REAL_TIME(formatAsString = false) {
 	let now = new Date();
 	return (!formatAsString) ? now.valueOf() : now.toLocaleString();
@@ -299,65 +164,7 @@ function UNIVERSE_TIME(formatAsString = false) {
 	return (!formatAsString) ? universeTime : new Date(universeTime).toUTCString();
 }
 
-function HOURS_TO_TIME_STRING(hours, includeSeconds = true, limitTo24Hours = true) {
-	if (hours < 0) return '- NEGATIVE -';
 
-	let h = hours;
-	let m = (h - Math.floor(h)) * 60;
-	let s = (m - Math.floor(m)) * 60;
-
-	s = Math.round(s);
-	if (s >= 60) {
-		s -= 60;
-		m += 1;
-	}
-
-	m = Math.floor(m);
-	if (m >= 60) {
-		m -= 60;
-		h += 1;
-	}
-
-	h = Math.floor(h);
-	if (limitTo24Hours) {
-		while (h >= 24) h -= 24;
-	}
-
-	let ampm = '';
-	if (limitTo24Hours && !window.SETTING_24HR) {
-		ampm = ' ' + GET_AM_PM(h);
-		h = CONVERT_24H_TO_12H(h);
-	}
-
-	h = (h < 10) ? '0' + h : h;
-	m = (m < 10) ? '0' + m : m;
-	s = (s < 10) ? '0' + s : s;
-
-	return h + ':' + m + (includeSeconds ? ':' + s : '') + ampm;
-}
-
-function DATE_TO_SHORT_TIME(date) {
-	let h = date.getHours();
-	let m = date.getMinutes();
-
-	let ampm = '';
-	if (!window.SETTING_24HR) {
-		ampm = ' ' + GET_AM_PM(h);
-		h = CONVERT_24H_TO_12H(h);
-	}
-
-	h = (h < 10) ? '0' + h : h;
-	m = (m < 10) ? '0' + m : m;
-
-	return h + ':' + m + ampm;
-}
-
-function GET_AM_PM(hour) { return (hour < 12) ? 'am' : 'pm'; }
-function CONVERT_24H_TO_12H(hour) {
-	if (hour > 12) hour -= 12;
-	if (hour === 0) hour = 12;
-	return hour;
-}
 
 
 function getSystemByName(string) {
@@ -411,6 +218,21 @@ async function startVerseTime() {
 }
 startVerseTime();
 
+function checkHash() {
+	let hash = window.location.hash;
+	if (hash === '') return;
+
+	let hashParts = hash.replace('#', '').replaceAll('_', ' ').split('@');
+
+	const locationName = hashParts[0];
+	const location = getLocationByName(locationName);
+
+	if (hashParts[1] !== undefined) {
+		setChosenTime(hashParts[1], true);
+	}
+
+	UI.setMapLocation(locationName);
+}
 
 window.addEventListener('hashchange', () => {
 	if (window.suppressReload) return;
